Refactor tests longer than 30s to run so that they no longer run when the race option is enabled, nor will they run if the -short flag is used.
